requirements.txt
streamlit>=1.38.0
pandas>=2.2.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
bcrypt
